## 🚀 Pull Request
Proposed revival of https://github.com/SciTools/iris/pull/3238

I was re-inspired to do this by a(nother) user issue with confused usage of a CubeList leading to mystifying error messages.
( like `mylist = CubeList([mylist, new_cube])` )

It foundered before because we kept thinking of new possible ways (methods) to add elements.
But I discovered a section in the Python docs, which I think is more complete + definitive
["Lists implement all of the common and mutable sequence operations.  Lists also provide the following additional methods"](https://docs.python.org/3/library/stdtypes.html#list)
So that makes me feel more secure about this now.

~Draft for now to check tests ...~

Ok now for real PR, I think