@rcomer 2022-05-25
That test failure is discussed here: https://github.com/SciTools/iris/pull/3238#discussion_r241853437.  Since we're no longer supporting python2, could just revert to using the `CubeList.copy()` method, rather than `copy.copy()`!