@pp-mo 2022-05-25
> That test failure is discussed here: [#3238 (comment)](https://github.com/SciTools/iris/pull/3238#discussion_r241853437). Since we're no longer supporting python2, could just revert to using the `CubeList.copy()` method, rather than `copy.copy()`!

Don't want to trip over each other, but..
I think that "__iadd__" should actually return the result
["do the operation in-place (modifying self) and return the result"](https://docs.python.org/3/reference/datamodel.html#object.__iadd__)
so I just fixed that.