@lbdreyer 2022-05-26
I have had a look and had a play and it all looks very reasonable to me; everything works as expected. 

I have also timed as @rcomer did in [this comment](https://github.com/SciTools/iris/issues/1897#issuecomment-427853668) and get similar performance difference with the` hasattr` as @rcomer got with `isinstance`

```
import iris.cube


def no_check(cubelist, cube):
    cubelist.append(cube)
    
def check(cubelist, cube):
    if not hasattr(cube, "add_aux_coord"):
        raise TypeError('Cubes only in a cubelist!')
        
        
mycube = iris.cube.Cube(1)
mycubelist = iris.cube.CubeList()
```

```
$ python -m timeit -r 10 -s 'import cubelist_append' "cubelist_append.no_check(cubelist_append.mycubelist, cubelist_append.mycube)"
1000000 loops, best of 10: 150 nsec per loop
$ python -m timeit -r 10 -s 'import cubelist_append' "cubelist_append.check(cubelist_append.mycubelist, cubelist_append.mycube)"
2000000 loops, best of 10: 161 nsec per loop
```

So unless someone is creating an enormous cubelist (which I suppose isn't out of the realms of possibility, especially if they were to use load_raw on a pp/grib file with many many fields) that should be an acceptably small slowdown.