@trexfeathers 2022-05-27
>So unless someone is creating an enormous cubelist (which I suppose isn't out of the realms of possibility, especially if they were to use load_raw on a pp/grib file with many many fields) that should be an acceptably small slowdown.

Our loading benchmarks include a 'deep' PP and FF, so I've just run them against this branch just in case it threw up something. We're all good:

`nox --session="benchmarks(branch)" -- upstream/main --bench=LoadAndRealise`

<details>
<summary>Expand for output</summary>

```
nox > asv compare caae4507 HEAD --factor=1.2 --split

Benchmarks that have stayed the same:

       before           after         ratio
     [caae4507]       [7a48fd66]
     <pr/4767~26>       <pr/4767> 
       11.9±0.4ms       11.8±0.2ms     0.99  load.LoadAndRealise.time_load((1280, 960, 5), False, 'FF')
         60.9±4ms         60.2±2ms     0.99  load.LoadAndRealise.time_load((1280, 960, 5), False, 'NetCDF')
       13.7±0.6ms       13.5±0.2ms     0.98  load.LoadAndRealise.time_load((1280, 960, 5), False, 'PP')
       11.9±0.5ms       12.0±0.3ms     1.01  load.LoadAndRealise.time_load((1280, 960, 5), True, 'FF')
         55.7±3ms         58.4±4ms     1.05  load.LoadAndRealise.time_load((1280, 960, 5), True, 'NetCDF')
       13.6±0.2ms       14.7±0.9ms     1.08  load.LoadAndRealise.time_load((1280, 960, 5), True, 'PP')
       1.70±0.02s       1.73±0.02s     1.02  load.LoadAndRealise.time_load((2, 2, 1000), False, 'FF')
         52.9±2ms         55.0±2ms     1.04  load.LoadAndRealise.time_load((2, 2, 1000), False, 'NetCDF')
       1.91±0.02s       1.92±0.01s     1.01  load.LoadAndRealise.time_load((2, 2, 1000), False, 'PP')
       1.71±0.02s       1.73±0.02s     1.01  load.LoadAndRealise.time_load((2, 2, 1000), True, 'FF')
         51.1±5ms         48.9±3ms     0.96  load.LoadAndRealise.time_load((2, 2, 1000), True, 'NetCDF')
       1.91±0.02s       1.92±0.02s     1.01  load.LoadAndRealise.time_load((2, 2, 1000), True, 'PP')
       5.80±0.3ms       6.00±0.4ms     1.03  load.LoadAndRealise.time_load((2, 2, 2), False, 'FF')
         46.1±2ms         48.4±1ms     1.05  load.LoadAndRealise.time_load((2, 2, 2), False, 'NetCDF')
      5.36±0.07ms       5.60±0.2ms     1.04  load.LoadAndRealise.time_load((2, 2, 2), False, 'PP')
       5.77±0.1ms       5.77±0.3ms     1.00  load.LoadAndRealise.time_load((2, 2, 2), True, 'FF')
         47.0±1ms         47.4±1ms     1.01  load.LoadAndRealise.time_load((2, 2, 2), True, 'NetCDF')
       5.39±0.2ms       5.60±0.1ms     1.04  load.LoadAndRealise.time_load((2, 2, 2), True, 'PP')
         82.6±3ms         88.6±1ms     1.07  load.LoadAndRealise.time_realise((1280, 960, 5), False, 'FF')
         69.5±2ms         69.3±1ms     1.00  load.LoadAndRealise.time_realise((1280, 960, 5), False, 'NetCDF')
         24.7±1ms         24.6±1ms     1.00  load.LoadAndRealise.time_realise((1280, 960, 5), False, 'PP')
         41.7±1ms         41.3±2ms     0.99  load.LoadAndRealise.time_realise((1280, 960, 5), True, 'FF')
          109±2ms          110±1ms     1.00  load.LoadAndRealise.time_realise((1280, 960, 5), True, 'NetCDF')
         40.6±2ms         39.6±2ms     0.98  load.LoadAndRealise.time_realise((1280, 960, 5), True, 'PP')
          545±8ms         555±10ms     1.02  load.LoadAndRealise.time_realise((2, 2, 1000), False, 'FF')
       6.05±0.2ms       6.53±0.7ms     1.08  load.LoadAndRealise.time_realise((2, 2, 1000), False, 'NetCDF')
          544±7ms         548±20ms     1.01  load.LoadAndRealise.time_realise((2, 2, 1000), False, 'PP')
         559±20ms          567±9ms     1.02  load.LoadAndRealise.time_realise((2, 2, 1000), True, 'FF')
       5.64±0.2ms       5.82±0.3ms     1.03  load.LoadAndRealise.time_realise((2, 2, 1000), True, 'NetCDF')
         552±10ms          560±5ms     1.02  load.LoadAndRealise.time_realise((2, 2, 1000), True, 'PP')
      1.47±0.04ms      1.43±0.04ms     0.98  load.LoadAndRealise.time_realise((2, 2, 2), False, 'FF')
       4.90±0.1ms       5.35±0.6ms     1.09  load.LoadAndRealise.time_realise((2, 2, 2), False, 'NetCDF')
      1.45±0.06ms      1.48±0.07ms     1.02  load.LoadAndRealise.time_realise((2, 2, 2), False, 'PP')
      1.49±0.04ms      1.48±0.05ms     0.99  load.LoadAndRealise.time_realise((2, 2, 2), True, 'FF')
       5.10±0.2ms       5.04±0.2ms     0.99  load.LoadAndRealise.time_realise((2, 2, 2), True, 'NetCDF')
      1.46±0.04ms      1.52±0.05ms     1.04  load.LoadAndRealise.time_realise((2, 2, 2), True, 'PP')

```

</details>